Move u32hasher into its own module, don't expose the Hasher for use.
